AbstractGame definition
AbstractGame is a mutual parent for all the game mods it can be played in. Expected product is the ability to play in Single-player mode ('Hangman' game type) and in Multi-player mode (where multiple agents play against themself). This abstract ancestor sets the basic protocol for all of it's descendants inheriting it's concrete methods and properties, but also implements their own abstract signatures.